Need assistance, new to coding.
Everything works in this program except the end bit, where i'm trying to have the user input dictate the color and price printed. I know what I have is wrong but I believe this is the direction i need to go. I am very new to coding so any input and help will be appreciated